refactor(shared): move formatStats helper to core
