+++ packages/core/src/client/formatStats.ts
