+++ packages/core/src/provider/core/createCompiler.ts
@@ -5,7 +5,6 @@ import {
   color,
   logger,
   prettyTime,
-  formatStats,
   TARGET_ID_MAP,
   type RspackConfig,
   type RspackCompiler,
@@ -15,7 +14,11 @@ import {
 import { initConfigs, type InitConfigsOptions } from './initConfigs';
 import type { InternalContext } from '../../types';
 import type { Stats, MultiStats, StatsCompilation } from '@rspack/core';
-import { isSatisfyRspackVersion, rspackMinVersion } from '../shared';
+import {
+  formatStats,
+  rspackMinVersion,
+  isSatisfyRspackVersion,
+} from '../shared';
 
 export async function createCompiler({
   context,
