+++ packages/shared/src/index.ts
@@ -9,7 +9,6 @@ export * from './mergeChainedOptions';
 export * from './mergeRsbuildConfig';
 export * from './types';
 export * from './config';
-export * from './format';
 export * from './chain';
 export * from './getLoaderOptions';
 export * from './patch';
